pyproject.toml
[project]
name = "langchain-products"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asttokens>=3.0.1",
    "groq>=1.0.0",
    "langchain>=1.2.8",
    "langchain-core>=1.2.8",
    "langchain-openai>=1.1.7",
    "openai>=2.16.0",
    "python-dotenv>=1.2.1",
]

[dependency-groups]
dev = [
    "ipykernel>=7.1.0",
]
